refactor(run): split the preflight messages out at the 500-line cap
`run.rs` reached 574 lines and CI refused it. The seam was already there:
everything from `LIVENESS_WAIT` down answers "why can this not run?" rather
than running it, and every message in it is built and asserted rather than
formatted inline.

`run/preflight.rs` takes `died_immediately`, `liveness_argv`,
`liveness_verdict` and `cache_miss` with their tests. `run.rs` drops to
344 lines and preflight is 245.

The split is mechanical — no behaviour change, and the tests moved with the
functions they cover, so what proves them is unchanged.

Workspace: 380 + 214 + 175 tests, clippy zero errors, rustfmt clean, no
file over the cap.
